# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * eFootball Tournament Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Player {
  id: number;
  name: string;
  createdAt: string;
}

export interface CreatePlayerRequest {
  name: string;
}

export type TournamentType =
  (typeof TournamentType)[keyof typeof TournamentType];

export const TournamentType = {
  league: "league",
  knockout: "knockout",
} as const;

export type TournamentStatus =
  (typeof TournamentStatus)[keyof typeof TournamentStatus];

export const TournamentStatus = {
  setup: "setup",
  active: "active",
  completed: "completed",
} as const;

export interface Tournament {
  id: number;
  name: string;
  type: TournamentType;
  status: TournamentStatus;
  createdAt: string;
}

export type CreateTournamentRequestType =
  (typeof CreateTournamentRequestType)[keyof typeof CreateTournamentRequestType];

export const CreateTournamentRequestType = {
  league: "league",
  knockout: "knockout",
} as const;

export interface CreateTournamentRequest {
  name: string;
  type: CreateTournamentRequestType;
  playerIds: number[];
}

export type TournamentDetailsType =
  (typeof TournamentDetailsType)[keyof typeof TournamentDetailsType];

export const TournamentDetailsType = {
  league: "league",
  knockout: "knockout",
} as const;

export type TournamentDetailsStatus =
  (typeof TournamentDetailsStatus)[keyof typeof TournamentDetailsStatus];

export const TournamentDetailsStatus = {
  setup: "setup",
  active: "active",
  completed: "completed",
} as const;

export interface TournamentDetails {
  id: number;
  name: string;
  type: TournamentDetailsType;
  status: TournamentDetailsStatus;
  createdAt: string;
  players: Player[];
  totalFixtures: number;
  completedFixtures: number;
}

export type StandingRowFormItem =
  (typeof StandingRowFormItem)[keyof typeof StandingRowFormItem];

export const StandingRowFormItem = {
  W: "W",
  D: "D",
  L: "L",
} as const;

export interface StandingRow {
  playerId: number;
  playerName: string;
  played: number;
  won: number;
  drawn: number;
  lost: number;
  goalsFor: number;
  goalsAgainst: number;
  goalDifference: number;
  points: number;
  form: StandingRowFormItem[];
}

export interface Fixture {
  id: number;
  tournamentId: number;
  round: number;
  homePlayerId: number;
  homePlayerName: string;
  awayPlayerId: number;
  awayPlayerName: string;
  homeScore: number | null;
  awayScore: number | null;
  played: boolean;
  knockoutPhase: string | null;
}

export interface SubmitResultRequest {
  homeScore: number;
  awayScore: number;
}

export type ListFixturesParams = {
  tournamentId: number;
  round?: number;
};