# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: eFootball Tournament Manager API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: players
    description: Player management
  - name: tournaments
    description: Tournament management
  - name: fixtures
    description: Fixture management
  - name: results
    description: Match result management
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create a new player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerRequest"
      responses:
        "201":
          description: Player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"

  /players/{id}:
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete a player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Player deleted

  /tournaments:
    get:
      operationId: listTournaments
      tags: [tournaments]
      summary: List all tournaments
      responses:
        "200":
          description: List of tournaments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tournament"
    post:
      operationId: createTournament
      tags: [tournaments]
      summary: Create a new tournament
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTournamentRequest"
      responses:
        "201":
          description: Tournament created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tournament"

  /tournaments/{id}:
    get:
      operationId: getTournament
      tags: [tournaments]
      summary: Get tournament details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tournament details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TournamentDetails"
        "404":
          description: Not found
    delete:
      operationId: deleteTournament
      tags: [tournaments]
      summary: Delete a tournament
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /tournaments/{id}/standings:
    get:
      operationId: getStandings
      tags: [tournaments]
      summary: Get league standings
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Standings table
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StandingRow"

  /fixtures:
    get:
      operationId: listFixtures
      tags: [fixtures]
      summary: List fixtures for a tournament
      parameters:
        - name: tournamentId
          in: query
          required: true
          schema:
            type: integer
        - name: round
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of fixtures
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fixture"

  /fixtures/{id}/result:
    put:
      operationId: submitResult
      tags: [results]
      summary: Submit or update match result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitResultRequest"
      responses:
        "200":
          description: Result saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]

    CreatePlayerRequest:
      type: object
      properties:
        name:
          type: string
      required: [name]

    Tournament:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [league, knockout]
        status:
          type: string
          enum: [setup, active, completed]
        createdAt:
          type: string
          format: date-time
      required: [id, name, type, status, createdAt]

    CreateTournamentRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [league, knockout]
        playerIds:
          type: array
          items:
            type: integer
      required: [name, type, playerIds]

    TournamentDetails:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [league, knockout]
        status:
          type: string
          enum: [setup, active, completed]
        createdAt:
          type: string
          format: date-time
        players:
          type: array
          items:
            $ref: "#/components/schemas/Player"
        totalFixtures:
          type: integer
        completedFixtures:
          type: integer
      required: [id, name, type, status, createdAt, players, totalFixtures, completedFixtures]

    StandingRow:
      type: object
      properties:
        playerId:
          type: integer
        playerName:
          type: string
        played:
          type: integer
        won:
          type: integer
        drawn:
          type: integer
        lost:
          type: integer
        goalsFor:
          type: integer
        goalsAgainst:
          type: integer
        goalDifference:
          type: integer
        points:
          type: integer
        form:
          type: array
          items:
            type: string
            enum: [W, D, L]
      required: [playerId, playerName, played, won, drawn, lost, goalsFor, goalsAgainst, goalDifference, points, form]

    Fixture:
      type: object
      properties:
        id:
          type: integer
        tournamentId:
          type: integer
        round:
          type: integer
        homePlayerId:
          type: integer
        homePlayerName:
          type: string
        awayPlayerId:
          type: integer
        awayPlayerName:
          type: string
        homeScore:
          type: integer
          nullable: true
        awayScore:
          type: integer
          nullable: true
        played:
          type: boolean
        knockoutPhase:
          type: string
          nullable: true
      required: [id, tournamentId, round, homePlayerId, homePlayerName, awayPlayerId, awayPlayerName, homeScore, awayScore, played, knockoutPhase]

    SubmitResultRequest:
      type: object
      properties:
        homeScore:
          type: integer
        awayScore:
          type: integer
      required: [homeScore, awayScore]