# lib/api-zod/src/generated/types/createTournamentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * eFootball Tournament Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateTournamentRequestType } from "./createTournamentRequestType";

export interface CreateTournamentRequest {
  name: string;
  type: CreateTournamentRequestType;
  playerIds: number[];
}