# lib/api-zod/src/generated/types/fixture.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * eFootball Tournament Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface Fixture {
  id: number;
  tournamentId: number;
  round: number;
  homePlayerId: number;
  homePlayerName: string;
  awayPlayerId: number;
  awayPlayerName: string;
  homeScore: number | null;
  awayScore: number | null;
  played: boolean;
  knockoutPhase: string | null;
}