# lib/api-zod/src/generated/types/standingRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * eFootball Tournament Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { StandingRowFormItem } from "./standingRowFormItem";

export interface StandingRow {
  playerId: number;
  playerName: string;
  played: number;
  won: number;
  drawn: number;
  lost: number;
  goalsFor: number;
  goalsAgainst: number;
  goalDifference: number;
  points: number;
  form: StandingRowFormItem[];
}