# lib/api-zod/src/generated/types/submitResultRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * eFootball Tournament Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface SubmitResultRequest {
  homeScore: number;
  awayScore: number;
}