# lib/api-zod/src/generated/types/tournament.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * eFootball Tournament Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { TournamentStatus } from "./tournamentStatus";
import type { TournamentType } from "./tournamentType";

export interface Tournament {
  id: number;
  name: string;
  type: TournamentType;
  status: TournamentStatus;
  createdAt: Date;
}