# lib/api-zod/src/generated/types/tournamentDetails.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * eFootball Tournament Manager API
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from "./player";
import type { TournamentDetailsStatus } from "./tournamentDetailsStatus";
import type { TournamentDetailsType } from "./tournamentDetailsType";

export interface TournamentDetails {
  id: number;
  name: string;
  type: TournamentDetailsType;
  status: TournamentDetailsStatus;
  createdAt: Date;
  players: Player[];
  totalFixtures: number;
  completedFixtures: number;
}